refactor: extract pose system to resolve elytra flying conflicts
