+++ src/main/java/ganymedes01/etfuturum/mixinplugin/EtFuturumEarlyMixins.java
@@ -13,6 +13,8 @@ import ganymedes01.etfuturum.configuration.configs.ConfigTweaks;
 import ganymedes01.etfuturum.core.utils.Logger;
 import ganymedes01.etfuturum.configuration.configs.ConfigWorld;
 import ganymedes01.etfuturum.lib.Reference;
+import ganymedes01.etfuturum.pose.PlayerPose;
+import ganymedes01.etfuturum.pose.PlayerPoseManager;
 import ganymedes01.etfuturum.swimming.SwimmingHooks;
 import net.minecraft.launchwrapper.Launch;
 import org.spongepowered.asm.mixin.MixinEnvironment;
@@ -115,6 +117,7 @@ public class EtFuturumEarlyMixins implements IFMLLoadingPlugin, IEarlyMixinLoade
 			mixins.add("closedrops.MixinEntityPlayerMP");
 		}
 
+		boolean enablePoseSystem = false;
 		if (ConfigMixins.enableElytra) {
 			mixins.add("backlytra.MixinEntityPlayer");
 			mixins.add("backlytra.MixinEntityLivingBase");
@@ -128,8 +131,8 @@ public class EtFuturumEarlyMixins implements IFMLLoadingPlugin, IEarlyMixinLoade
 				mixins.add("backlytra.client.MixinEntityPlayerSP");
 				mixins.add("backlytra.client.MixinRenderPlayer");
 				mixins.add("backlytra.client.MixinModelBiped");
-				mixins.add("backlytra.client.MixinEntityRenderer");
 			}
+			enablePoseSystem = true;
 		}
 
 		if (ConfigMixins.enableDoWeatherCycle) {
@@ -276,18 +279,28 @@ public class EtFuturumEarlyMixins implements IFMLLoadingPlugin, IEarlyMixinLoade
 			mixins.add("swimming.MixinEntityLivingBase");
 			mixins.add("swimming.MixinEntityPlayer");
 			if (side == MixinEnvironment.Side.CLIENT) {
-				mixins.add("swimming.client.C04PacketPlayerPositionMixin");
-				mixins.add("swimming.client.C06PacketPlayerPosLookMixin");
 				mixins.add("swimming.client.MixinEntityClientPlayerMP");
 				mixins.add("swimming.client.MixinEntityPlayerSP");
-				mixins.add("swimming.client.MixinEntityRenderer");
 				mixins.add("swimming.client.MixinModelBiped");
 				mixins.add("swimming.client.MixinPlayerControllerMP");
 				mixins.add("swimming.client.MixinRenderPlayer");
-				mixins.add("swimming.client.NetHandlerPlayClientMixin");
 			}
+			enablePoseSystem = true;
 		} else if (ConfigMixins.enableModernSwimming) {
 			Logger.warn("Modern swimming mixins are disabled because swimmingDataWatcherFlag is reserved or conflicts with elytraDataWatcherFlag.");
+		}
+
+		if (enablePoseSystem)
+		{
+			mixins.add("pose.MixinEntity");
+			mixins.add("pose.MixinEntityPlayer");
+			if (side == MixinEnvironment.Side.CLIENT) {
+				mixins.add("pose.client.C04PacketPlayerPositionMixin");
+				mixins.add("pose.client.C06PacketPlayerPosLookMixin");
+				mixins.add("pose.client.MixinEntityClientPlayerMP");
+				mixins.add("pose.client.MixinEntityRenderer");
+				mixins.add("pose.client.NetHandlerPlayClientMixin");
+			}
 		}
 
 		if (false) { //Does not work for some reason, investigate in 2.6.1
